client/requirements.txt
requests==2.26.0
getmac==0.8.2
python-dotenv==0.19.0
geocoder==1.38.1
scapy==2.4.5
